seminar3 Ex2.2

+++ Семинар3/Example002/Program.cs
@@ -18,5 +18,5 @@ for (int i=0; i<N; i++)
 {
     sum += array[i];
 }
-double mean = sum / N;
+double mean = sum / (double)N;
 Console.WriteLine(mean);
